# chore: Optimize BuildTracestate()