Optimize `DistributedTracePayloadHandler.BuildTracestate()`

Benchmarking results: 
```
| Method                             | Mean     | Error    | StdDev   | Gen0   | Allocated |
|----------------------------------- |---------:|---------:|---------:|-------:|----------:|
| BuildTracestateOriginal_Benchmark  | 40.12 us | 0.771 us | 1.771 us | 6.5918 |  86.17 KB |
| BuildTracestateOptimized_Benchmark | 23.31 us | 0.455 us | 0.695 us | 4.5166 |  57.95 KB |
```

Also includes a few minor optimizations in `AgentBridge.cs`.